update: change the defintion of files add: http example for file-upload #11

+++ src/core/pipepress.ts
@@ -124,15 +124,17 @@ export class PipePress extends Router {
                         /* loop over files and delete temp files */
                         for (const filedName in ctx.files) {
                             const files = ctx.files[filedName];
-                            for (const file of files) {
-                                /* async delete the file */
-                                import('fs').then(fs => {
-                                    fs.unlink(file.path, (err) => {
-                                        if (err) {
-                                            console.error(`Could not delete temp file: ${file.path}`);  // TODO: how to handle this properly?
-                                        }
+                            if (files) {
+                                for (const file of files) {
+                                    /* async delete the file */
+                                    import('fs').then(fs => {
+                                        fs.unlink(file.path, (err) => {
+                                            if (err) {
+                                                console.error(`Could not delete temp file: ${file.path}`);  // TODO: how to handle this properly?
+                                            }
+                                        });
                                     });
-                                });
+                                }
                             }
                         }
                     }
@@ -248,7 +250,7 @@ export class PipePress extends Router {
             params,
             query: queryParams,
             body: undefined,
-            files: {}
+            files: undefined
         };
         return ctx;
     }
